requirements.txt
Django==4.0.1
python-dotenv==0.20.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
python-social-auth==0.3.6
psycopg2==2.9.3
gunicorn==20.1.0